refactor: BookShelfPreview main 컴포넌트 폴더로 이동
